split out tests, add status parsing test
