+++ mongocrypt/Cargo.toml
@@ -14,3 +14,4 @@ repository = "https://github.com/mongodb/libmongocrypt-rust"
 
 [dependencies]
 mongocrypt-sys = { path = "../mongocrypt-sys" }
+bson = "2.3.0"
